LA-25 moving app.ts and index.ts to src dir
