+++ backEnd/src/controllers/company.controller.ts
@@ -3,7 +3,7 @@ import { Request, Response } from 'express';
 
 
 export async function getAllProfileHandler(req: Request, res: Response) {
-  const profile = await getAllProfileService({ projectId });
+  const profile = await getAllProfilesService({ projectId });
   
   return res.send(profile);
 }
